requirements.txt
feedparser>=6.0.10
requests>=2.31.0
python-dateutil>=2.8.2
openai==0.28.0
